+++ src/earthkit/geo/regrid/array.py
@@ -11,7 +11,7 @@
 
 def regrid(data, in_grid=None, out_grid=None, *, interpolation="linear", backend="mir", **kwargs):
     r"""Array interface."""
-    from earthkit.geo.grids.regrid.data.numpy import handler
+    from earthkit.geo.regrid.data.numpy import handler
 
     h = handler()
     kwargs = kwargs.copy()
