+++ src/earthkit/geo/regrid/backends/precomputed.py
@@ -20,7 +20,7 @@ class MatrixBackend(Backend):
         self.db = self.get_db(inventory)
 
     def regrid(self, data, in_grid, out_grid, interpolation):
-        from earthkit.geo.grids.regrid.gridspec import _GridSpec
+        from earthkit.geo.regrid.gridspec import _GridSpec
 
         _out_grid = _GridSpec.from_any(out_grid)
