+++ src/earthkit/geo/regrid/data/numpy.py
@@ -17,7 +17,7 @@ LOG = logging.getLogger(__name__)
 class NumpyDataHandler(DataHandler):
     @staticmethod
     def match(data):
-        import earthkit.geo.grids.regrid.data.numpy as np
+        import numpy as np
 
         return isinstance(data, np.ndarray)
