+++ tests/regrid/regrid_backends/test_matrix_fieldlist.py
@@ -11,7 +11,7 @@ import numpy as np
 import pytest
 from eckit.geo import Grid
 
-from earthkit.geo import regrid
+from earthkit.geo.regrid import regrid
 from earthkit.geo.utils.testing import (
     NO_EKD,  # noqa: E402
     SYSTEM_MATRIX_BACKEND_NAME,  # noqa: E402
