+++ tests/regrid/regrid_backends/test_matrix_xarray.py
@@ -9,7 +9,7 @@
 
 import pytest
 
-from earthkit.geo import regrid
+from earthkit.geo.regrid import regrid
 from earthkit.geo.utils.testing import (
     NO_EKD,  # noqa: E402
     NO_MIR,  # noqa: E402
